@@ -44,23 +44,23 @@
     "@nuxtjs/robots": "^2.5.0",
     "@nuxtjs/sitemap": "^2.4.0",
     "@octokit/rest": "~21.0.2",
-    "a11y-dialog": "^8.1.0",
+    "a11y-dialog": "^8.1.1",
     "ajv": "^8.17.1",
     "ajv-formats": "^3.0.1",
     "chalk": "^5.3.0",
     "color-hash": "^2.0.2",
-    "color-name-list": "^10.25.1",
+    "color-name-list": "^10.28.0",
     "cookie-universal-nuxt": "^2.2.2",
     "cors": "^2.8.5",
     "embetty-vue": "^2.0.1",
-    "express": "^4.19.2",
+    "express": "^4.21.1",
     "jszip": "^3.10.1",
     "node-env-file": "^0.1.8",
     "nuxt": "~2.18.1",
     "nuxt-helmet": "^3.0.0",
-    "openapi-backend": "^5.10.6",
+    "openapi-backend": "^5.11.0",
     "sanitize-filename": "^1.6.3",
-    "sass": "^1.77.8",
+    "sass": "^1.79.4",
     "sass-loader": "^10.5.2",
     "scroll-into-view": "^1.16.2",
     "svg-inline-loader": "^0.8.2",
@@ -72,28 +72,31 @@
     "xmlbuilder": "^15.1.1"
   },
   "devDependencies": {
+    "@eslint/compat": "^1.2.0",
+    "@eslint/js": "~9.12.0",
+    "@eslint/markdown": "~6.2.0",
     "@stylistic/stylelint-config": "~2.0.0",
     "broken-link-checker": "^0.7.8",
     "dir-compare": "^5.0.0",
     "disparity": "^3.2.0",
-    "eslint": "~8.57.0",
-    "eslint-plugin-import": "~2.29.1",
+    "eslint": "~9.12.0",
+    "eslint-plugin-import": "~2.31.0",
     "eslint-plugin-jsdoc": "~50.0.1",
     "eslint-plugin-jsonc": "~2.16.0",
-    "eslint-plugin-markdown": "~5.1.0",
     "eslint-plugin-nuxt": "~4.0.0",
     "eslint-plugin-promise": "~7.1.0",
     "eslint-plugin-sonarjs": "~1.0.4",
-    "eslint-plugin-unicorn": "~55.0.0",
+    "eslint-plugin-unicorn": "~56.0.0",
     "eslint-plugin-vue": "~9.28.0",
     "eslint-plugin-vuejs-accessibility": "~2.4.1",
-    "jsdoc-to-markdown": "^9.0.0",
+    "globals": "^15.10.0",
+    "jsdoc-to-markdown": "^9.0.2",
     "libxmljs": "^1.0.11",
     "minimist": "^1.2.8",
     "postcss-html": "^1.7.0",
     "stylelint": "~16.9.0",
     "stylelint-config-html": "~1.1.0",
-    "stylelint-config-recess-order": "~5.1.0",
+    "stylelint-config-recess-order": "~5.1.1",
     "stylelint-config-standard-scss": "~13.1.0",
     "terminate": "^2.6.1",
     "widdershins": "^4.0.1"
