@@ -1,11 +1,11 @@
 import {
-  manufacturersSchema,
+  capabilitySchema,
+  channelSchema,
+  definitionsSchema,
   fixtureRedirectSchema,
   fixtureSchema,
-  channelSchema,
-  capabilitySchema,
+  manufacturersSchema,
   wheelSlotSchema,
-  definitionsSchema,
 } from '../lib/esm-shim.cjs';
 
 export const fixtureProperties = fixtureSchema.properties;
