@@ -2,9 +2,9 @@ import xml2js from 'xml2js';
 
 import importJson from '../../lib/import-json.js';
 import {
-  getCapabilityFromChannelPreset,
-  getCapabilityFromCapabilityPreset,
   capabilityPresets,
+  getCapabilityFromCapabilityPreset,
+  getCapabilityFromChannelPreset,
   importHelpers,
 } from './presets.js';
 
