@@ -5,8 +5,8 @@ import JSZip from 'jszip';
 import importJson from '../../lib/import-json.js';
 import Fixture from '../../lib/model/Fixture.js';
 import Manufacturer from '../../lib/model/Manufacturer.js';
-import { fixtureFromRepository, embedResourcesIntoFixtureJson } from '../../lib/model.js';
-import { sendJson, sendAttachment } from '../../lib/server-response-helpers.js';
+import { embedResourcesIntoFixtureJson, fixtureFromRepository } from '../../lib/model.js';
+import { sendAttachment, sendJson } from '../../lib/server-response-helpers.js';
 /** @typedef {import('http').ServerResponse} ServerResponse */
 
 const pluginsPromise = importJson(`../../plugins/plugins.json`, import.meta.url);
