@@ -246,12 +246,12 @@ import { objectProp } from 'vue-ts-types';
 import { capabilityTypes, channelProperties } from '../../../lib/schema-properties.js';
 import {
   constants,
-  getEmptyFormState,
   getEmptyCapability,
   getEmptyFineChannel,
+  getEmptyFormState,
   getSanitizedChannel,
-  isChannelChanged,
   isCapabilityChanged,
+  isChannelChanged,
 } from '../../assets/scripts/editor-utils.js';
 
 import A11yDialog from '../A11yDialog.vue';
