@@ -203,7 +203,7 @@
 import { numberProp, objectProp } from 'vue-ts-types';
 import Draggable from 'vuedraggable';
 
-import { schemaDefinitions, modeProperties } from '../../../lib/schema-properties.js';
+import { modeProperties, schemaDefinitions } from '../../../lib/schema-properties.js';
 import { constants } from '../../assets/scripts/editor-utils.js';
 
 import LabeledInput from '../LabeledInput.vue';
