@@ -48,7 +48,7 @@
 
 <script>
 import { objectProp } from 'vue-ts-types';
-import { schemaDefinitions, entitiesSchema } from '../../../../lib/schema-properties.js';
+import { entitiesSchema, schemaDefinitions } from '../../../../lib/schema-properties.js';
 import { colorsHexStringToArray } from '../../../assets/scripts/editor-utils.js';
 
 import LabeledInput from '../../LabeledInput.vue';
